Move plugin OAuth helpers into plugins service
Interfaces:

- services/plugins/oauth.UpstreamConfig configures OAuth authorization and token exchange for plugin auth handlers.

- services/plugins/oauth.NewUpstream(cfg, opts...) returns an UpstreamHandler with AuthorizationURL, AuthorizationURLWithPKCE, ExchangeCode, and RefreshToken methods.

- services/plugins/oauth.ManualAuthHandler preserves manual-token auth behavior for declarative providers.

- services/plugins/oauth.GenerateVerifier and ComputeS256Challenge continue to expose PKCE helpers for OAuth and MCP authorization flows.

Examples:

handler := oauth.NewUpstream(oauth.UpstreamConfig{ClientID: clientID, TokenURL: tokenURL, RedirectURL: redirectURL}, oauth.WithHTTPClient(httpClient))

token, err := handler.ExchangeCode(ctx, code)

manual := oauth.ManualAuthHandler{}; token, err = manual.ExchangeCode(ctx, submittedToken)
