+++ gestaltd/internal/mcpoauth/handler.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/valon-technologies/gestalt/server/core"
-	"github.com/valon-technologies/gestalt/server/internal/oauth"
+	"github.com/valon-technologies/gestalt/server/services/plugins/oauth"
 )
 
 const discoveryCacheTTL = 24 * time.Hour
