+++ gestaltd/internal/server/handlers_mcp_authorization_server.go
@@ -12,8 +12,8 @@ import (
 	"strings"
 
 	"github.com/valon-technologies/gestalt/server/core"
-	"github.com/valon-technologies/gestalt/server/internal/oauth"
 	"github.com/valon-technologies/gestalt/server/services/identity/principal"
+	"github.com/valon-technologies/gestalt/server/services/plugins/oauth"
 )
 
 const (
