+++ gestaltd/services/plugins/declarative/base.go
@@ -10,9 +10,9 @@ import (
 
 	"github.com/valon-technologies/gestalt/server/core"
 	"github.com/valon-technologies/gestalt/server/core/catalog"
-	"github.com/valon-technologies/gestalt/server/internal/oauth"
 	"github.com/valon-technologies/gestalt/server/services/egress"
 	"github.com/valon-technologies/gestalt/server/services/plugins/apiexec"
+	"github.com/valon-technologies/gestalt/server/services/plugins/oauth"
 	"github.com/valon-technologies/gestalt/server/services/plugins/paraminterp"
 )
