+++ gestaltd/services/plugins/oauth/manual.go
